Major refractor taking Pierre feedback/pair programming experience
