--- src/main/resources/hudson/plugins/audit_trail/LogFileAuditLogger/help-rotateDaily.html
@@ -1,6 +1,0 @@
-<div>
-    Rotate the logs daily according to the time configured in the System. The FileHandler will be configured with
-    limit=0 (unlimited size) and count=1 so that it does not produce more than 1 file per day. Thus the value
-    configured in <i>%Log File Size MB</i> will not be considered and the <i>%Log file Count</i> will be used to
-    limit the number of files created in the file system.
-</div>
