+++ src/main/resources/hudson/plugins/audit_trail/LogFileDailyRotationAuditLogger/help-count.html
@@ -1,0 +1,4 @@
+<div>
+  Number of rotating log files to use (see
+  <a href="https://docs.oracle.com/javase/8/docs/api/java/util/logging/FileHandler.html#FileHandler-java.lang.String-int-int-">FileHandler</a>).
+</div>
