+++ src/main/resources/hudson/plugins/audit_trail/LogFileDailyRotationAuditLogger/help-log.html
@@ -1,0 +1,8 @@
+<div>
+  Full path for log location and filename. May contain <a
+  href="https://docs.oracle.com/javase/8/docs/api/java/util/logging/FileHandler.html"
+  >FileHandler</a> pattern tokens, such as <tt>%g</tt> for a number
+  to distinguish rotated logs.
+  Environment variables will be expanded, eg ${JENKINS_HOME} will be expanded provided
+  the JENKINS_HOME environment variable is defined on the controller.
+</div>
