+++ src/main/resources/hudson/plugins/audit_trail/LogFileDailyRotationAuditLogger/help-logSeparator.html
@@ -1,0 +1,4 @@
+<div>
+    A separator to set between the date timestamp and the logged message fields to ease parsing of the file.
+    Default is a space.
+</div>
