+++ src/test/java/hudson/plugins/audit_trail/SimpleAuditTrailPluginConfiguratorHelper.java
@@ -16,7 +16,6 @@ public class SimpleAuditTrailPluginConfiguratorHelper {
     private static final String LOG_LOCATION_INPUT_NAME = "_.log";
     private static final String LOG_FILE_SIZE_INPUT_NAME = "_.limit";
     private static final String LOG_FILE_COUNT_INPUT_NAME = "_.count";
-    private static final String LOG_FILE_ROTATE_DAILY= "_.rotateDaily";
     private static final String LOG_FILE_LOG_SEPARATOR_INPUT_NAME = "_.logSeparator";
     private static final String PATTERN_INPUT_NAME= "pattern";
     private static final String LOG_BUILD_CAUSE_INPUT_NAME="logBuildCause";
@@ -57,24 +56,6 @@ public class SimpleAuditTrailPluginConfiguratorHelper {
         form.getInputByName(LOG_LOCATION_INPUT_NAME).setValueAttribute(logFile.getPath());
         form.getInputByName(LOG_FILE_SIZE_INPUT_NAME).setValueAttribute("1");
         form.getInputByName(LOG_FILE_COUNT_INPUT_NAME).setValueAttribute("2");
-        form.getInputByName(LOG_FILE_ROTATE_DAILY).setChecked(false);
-        form.getInputByName(LOG_FILE_LOG_SEPARATOR_INPUT_NAME).setValueAttribute(DEFAULT_LOG_SEPARATOR);
-        form.getInputByName(PATTERN_INPUT_NAME).setValueAttribute(pattern);
-        form.getInputByName(LOG_BUILD_CAUSE_INPUT_NAME).setChecked(logBuildCause);
-        form.getInputByName(LOG_CREDENTIALS_USAGE_INPUT_NAME).setChecked(logCredentialsUsage);
-        j.submit(form);
-    }
-
-    public void sendConfigurationToRotateDaily(JenkinsRule j, JenkinsRule.WebClient wc) throws Exception {
-        HtmlPage configure = wc.goTo("configure");
-        HtmlForm form = configure.getFormByName("config");
-        j.getButtonByCaption(form, ADD_LOGGER_BUTTON_TEXT).click();
-        configure.getAnchorByText(LOG_FILE_COMBO_TEXT).click();
-        wc.waitForBackgroundJavaScript(TIMEOUT);
-        form.getInputByName(LOG_LOCATION_INPUT_NAME).setValueAttribute(logFile.getPath());
-        form.getInputByName(LOG_FILE_SIZE_INPUT_NAME).setValueAttribute("0");
-        form.getInputByName(LOG_FILE_COUNT_INPUT_NAME).setValueAttribute("5");
-        form.getInputByName(LOG_FILE_ROTATE_DAILY).setChecked(true);
         form.getInputByName(LOG_FILE_LOG_SEPARATOR_INPUT_NAME).setValueAttribute(DEFAULT_LOG_SEPARATOR);
         form.getInputByName(PATTERN_INPUT_NAME).setValueAttribute(pattern);
         form.getInputByName(LOG_BUILD_CAUSE_INPUT_NAME).setChecked(logBuildCause);
