+++ src/test/resources/hudson/plugins/audit_trail/expected.yml
@@ -10,7 +10,6 @@ loggers:
     limit: 666
     log: "/log/location"
     logSeparator: ";"
-    rotateDaily: false
 - syslog:
     appName: "jenkins"
     facility: "DAEMON"
