+++ src/test/resources/hudson/plugins/audit_trail/sample.xml
@@ -6,7 +6,6 @@
     <logBuildCause>true</logBuildCause>
     <loggers>
         <hudson.plugins.audit__trail.LogFileAuditLogger>
-            <rotateDaily>false</rotateDaily>
             <log>/tmp/xml-logs</log>
             <limit>100</limit>
             <count>5</count>
